Fixes property content serialization detection when using a property named `Item`. (#5996)
Fix property content serialization detection when using a property named `Item`.
